requirements.txt
hydra-core
psutil
matplotlib
annoy
hnswlib
scann
pynndescent
tensorflow
numpy
